pyproject.toml
[project]
name = "daiv"
version = "2.0.0"
description = """\
  Async SWE agents seamlessly integrated on your git platform to automate code issues implementation, reviews, and \
  pipeline repairs.\
  """
readme = "README.md"
license = { file = "LICENSE" }
maintainers = [
  { name = "Sandro Rodrigues", email = "srtabs@gmail.com" },
]
authors = [
  { name = "Sandro Rodrigues", email = "srtabs@gmail.com" },
]
requires-python = ">=3.14,<3.15"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.14",
]
dependencies = [
  "ag-ui-langgraph==0.0.31",
  "copilotkit==0.1.86",
  "croniter==6.2.2",
  "ddgs==9.13.0",
  "deepagents==0.4.12",
  "django==6.0.4",
  "django-allauth[socialaccount]==65.15.1",
  "django-crontask==1.1.3",
  "django-extensions==4.1.0",
  "django-ninja==1.6.2",
  "django-oauth-toolkit==3.2.0",
  "django-split-settings==1.3.2",
  "django-tasks==0.12.0",
  "django-tasks-db==0.12.0",
  "get-docker-secret==2.0.0",
  "gitpython==3.1.46",
  "httpx==0.28.1",
  "ipython==9.12.0",
  "jinja2==3.1.6",
  "langchain[anthropic,community,google-genai,openai]==1.2.15",
  "langchain-mcp-adapters==0.2.2",
  "langgraph==1.1.6",
  "langgraph-checkpoint-redis==0.4.0",
  "langsmith[pytest]==0.7.26",
  "langsmith-fetch==0.3.1",
  "markdownify==1.2.2",
  "mcp[cli]==1.27.0",
  "parable",
  "prompt-toolkit==3.0.52",
  "psycopg[pool]==3.3.3",
  "pydantic==2.12.5",
  "pydantic-settings==2.13.1",
  "pygithub==2.9.0",
  "python-decouple==3.8.0",
  "python-gitlab==8.2.0",
  "pyyaml==6.0.3",
  "redis==7.4.0",
  "sentry-sdk==2.57.0",
  "unidiff==0.7.5",
  "uvicorn[standard]==0.44.0",
  "whitenoise==6.12.0",
]
urls.changelog = "https://github.com/srtab/daiv/blob/main/CHANGELOG.md"
urls.issues = "https://github.com/srtab/daiv/issues"
urls.source = "https://github.com/srtab/daiv"

[dependency-groups]
dev = [
  "coverage==7.13.5",
  "datasets==4.8.4",
  "openevals==0.2.0",
  "prek==0.3.8",
  "pyproject-fmt==2.21.0",
  "pytest==9.0.3",
  "pytest-asyncio==1.3.0",
  "pytest-cov==7.1.0",
  "pytest-django==4.12.0",
  "pytest-env==1.6.0",
  "pytest-httpx==0.36.0",
  "pytest-mock==3.15.1",
  "pytest-xdist==3.8.0",
  "python-dotenv==1.2.2",
  "ruff==0.15.9",
  "ty==0.0.29",
  "types-pyyaml==6.0.12.20250915",
  "watchdog==6.0.0",
]
docs = [
  "mike==2.1.4",
  "mkdocs==1.6.1",
  "mkdocs-git-revision-date-localized-plugin==1.5.1",
  "mkdocs-llmstxt==0.5.0",
  "mkdocs-material==9.7.6",
]

[tool.uv]
default-groups = [ "docs", "dev" ]
sources.parable = { git = "https://github.com/ldayton/Parable.git", rev = "9f7e15734bd3e5bb9fe252d2491d6f61572f0e57" }

[tool.ruff]
line-length = 120
extend-exclude = [ "*/node_modules/*", "*/static/*" ]
force-exclude = true
preview = true
unsafe-fixes = true
show-fixes = true
format.skip-magic-trailing-comma = true
lint.select = [
  # flake8-builtins
  "A",
  # flake8-bugbear
  "B",
  # flake8-comprehensions
  "C4",
  # flake8-django
  "DJ",
  # pycodestyle
  "E",
  # eradicate
  "ERA",
  # Pyflakes
  "F",
  # flake8-logging-format
  "G",
  # isort
  "I",
  # flake8-logging
  "LOG",
  # pep8-naming
  "N",
  # flake8-use-pathlib
  "PTH",
  # flake8-bandit
  "S",
  # flake8-simplify
  "SIM",
  # flake8-print
  "T20",
  # flake8-type-checking
  "TCH",
  # pyupgrade
  "UP",
  # pycodestyle
  "W",
]
lint.ignore = [ "N812", "S101" ]
lint.per-file-ignores."*/migrations/*" = [ "E501" ]
lint.isort.known-first-party = [
  "accounts",
  "automation",
  "chat",
  "codebase",
  "core",
  "schedules",
  "slash_commands",
  "daiv",
]
lint.isort.section-order = [
  "future",
  "standard-library",
  "django",
  "third-party",
  "first-party",
  "local-folder",
]
lint.isort.sections.django = [ "django" ]
lint.isort.split-on-trailing-comma = false

[tool.pyproject-fmt]
keep_full_version = true

[tool.ty]
environment.python-version = "3.14"
src.exclude = [ "**/tests/**", "**/docs/**" ]
rules.unresolved-import = "ignore"

[tool.pytest]
ini_options.DJANGO_SETTINGS_MODULE = "daiv.settings.test"
ini_options.pythonpath = "daiv"
ini_options.python_files = "test_*.py"
ini_options.python_classes = "Test *Test"
ini_options.testpaths = [ "tests" ]
ini_options.norecursedirs = [
  ".venv",
  "data",
  "docker",
  "docs",
  "evals",
  "node_modules",
  "static",
]
ini_options.addopts = "--cov --cov-config=.coveragerc -p no:warnings"
ini_options.asyncio_mode = "auto"
ini_options.asyncio_default_fixture_loop_scope = "function"

[tool.pytest_env]
DJANGO_SETTINGS_MODULE = "daiv.settings.test"
DJANGO_SECRET_KEY = "not-so-secret"
ANTHROPIC_API_KEY = "anthropic-api-key"
OPENAI_API_KEY = "openai-api-key"
CODEBASE_GITLAB_WEBHOOK_SECRET = "gitlab-webhook-secret"
CODEBASE_GITHUB_INSTALLATION_ID = "123456"
NINJA_SKIP_REGISTRY = true
